pyproject.toml
[project]
name = "fastFigma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-fasthtml>=0.12.19",
    "requests>=2.32.4",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["fastFigma*"]